Worked on understanding velocity formula by Kayser
Added the derivation that I previously done in FYP which gives an identical results to that by Kayser. Also updated progress log and included further direction, including the speculation that the modification in lens velocity (microlens) is actually due to the fact that the microlens itself is lensed by the background matter.